cloud_fonction_prediction_prescription_maladie/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-iot